+++ NEWS.md
@@ -1,7 +1,7 @@
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-# eq5d (development version)
+# eq5d 0.15.4 (2024-09-04)
 
 - Trinidad and Tobago EQ-5D-5L value set added.